Move CleanupEnvironment function to internal/env
This function was duplicated across databricks/, databricks/client/,
and internal/ (for acceptance tests). It should only be used for
testing and as such is best placed under internal/ to prevent it
being part of our public API surface.

Instead of using defer for cleanup it now hooks into t.Cleanup.