Fixes for the repos integration test (#110)
There were two issues with this test:

1. It didn't use the synthesized root path. This meant that if another
test failed to clean up after itself, a subsequent run would fail
because the path already existed. In this change we use a random string
for the checkout path.

2. If "Files in Repos" isn't enabled in the workspace, it would fail
with "NOTE: Databricks Repos will only clone Databricks notebooks from
the remote repo. This repo does not contain any Databricks notebooks.".
We now check if "Files in Repos" is enabled and skip the test if so.