Implement [io.WriterTo] for DBFS file reader
I noticed ~30 API calls to download a 1.44M file from DBFS and it
turns out that [io.ReadAll] starts out with a small buffer and
grows it conservatively.

Through the [io.WriterTo] interface we can make the same reader
use the maximum DBFS read size of 1MB and reduce the number of
API calls for the same file down to 2.

This PR also contains a related fix for [FileReader.Read] such that
it no longer requires a final zero-length read call.