Make logger context aware (#333)
This commit changes a few things:
* The logging functions in the `logger` package now take a context
argument. A logger from the context is used if one is configured,
otherwise these functions use the global `DefaultLogger`.
* Add functions `NewContext` to configure a context with a logger and
`FromContext` to retrieve a logger from a context.
* The `logger.Logger` interface defines an `Enabled` function to check
if logging at a level is enabled or not. This means we no longer compute
a redacted copy of the response body for every response if debug logging
is not enabled.
* An example under `examples/slog` that shows how to use the SDK with
slog.
* An example under `examples/zerolog` that shows how to use the SDK with
zerolog.