Fix cleanup logic in SQL warehouse integration test (#400)
## Changes

This was changed in #389 without testing.

The function `w.Warehouses.DeleteByIdAndWait` always fails because upon
deletion there is nothing left to GET and it returns a 404 error. For
the sake of test cleanup, we can issue the delete and move on without
waiting.

## Tests

Ran TestAccSqlWarehouses manually to confirm this fixes the issue.