Parse Camel Case and Pascal Case Enum Values (#550)
## Changes
Currently, enum values that use camelCase or PascalCase are not split
into separate words during the parse phase of the OpenAPI spec. As a
result, `PascalCase()` returns `Camelcase` or `Pascalcase`,
respectively. To fix this, I've modified SplitASCII to support parsing
more forms (snake-case and constant-case), to split words based on all
non-alphanumeric symbols, and to estimate the "case" of a number by
looking at the neighboring letters.

## Tests
- [x] Added several cases to SplitASCII unit tests.
- [x] Regenerated the Go, Python, and Java SDKs with this change, and
the results are positive.
https://github.com/databricks/databricks-sdk-py/pull/229 and
https://github.com/databricks/databricks-sdk-java/pull/115