Fix names of keyword arguments in examples (#560)
## Changes
When generating the names of the arguments to use in methods named
XByYAndZ (e.g. DeleteByJobId), the field names are derived from the name
of the method. The logic for this was incorrect. This doesn't affect Go,
since Go examples don't use the `.Request()` method on `call` but
instead read from the call name and args directly.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied