Improve semantics for listing methods
This PR moves listing requests to an `ListXXXInternal` entity, which keeps all the low-level pagination request fields, like `offset`, `limit`, `token`, `page`. Original entity keeps all of the fields, except pagination-related. This allows to mitigate a common confusion, where users supply `limit` for the iterator and it acting only a _result page limit_, not the _max number of results_.