Support Files API (application/octet-stream) in OpenAPI Spec (#572)
## Changes
This PR adds support for non-application/json request and response
content types in the Databricks OpenAPI specification. This PR
introduces the following changes to the OpenAPI library interface:

**Entity**
* `IsByteStream bool` -> `true` iff the entity should be modeled as a
byte stream. This is only allowed for top-level entities (i.e. that are
used directly as a request or response) or for the field within such an
entity, provided that that field is the "contents" field (for requests)

**Method**
* `RequestBodyField *Field` -> a pointer to the field of the request
entity that is the body of the request. Only to be used when the request
body can be captured by a single field (today, this is only allowed for
the "contents" field of non-application/json requests).
* `FixedRequestHeaders map[string]string` -> the headers which are
always the same, regardless of request parameters. In the future, a new
method could be added to support fields from requests that are provided
as headers.
* `(m *Method) IsRequestByteStream() bool` -> `true` iff the request is
modeled as a byte stream (i.e. != application/json)
* `(m *Method) IsResponseByteStream() bool` -> `true` iff the response
is modeled as a byte stream (i.e. != application/json).

**Parameter**
* `MultiSegment bool `json:"x-databricks-multi-segment,omitempty"`` ->
indicates that this parameter is a path parameter that may contain
slashes, which is not officially compliant with OpenAPI. (unused today)

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied