Propagate Request Headers to client.go (#589)
## Changes
Based on changes in
https://github.com/databricks/databricks-sdk-go/pull/572.

This PR adds initial support for headers in client.go. The `impl.go`
implementations, when making a request, will construct a map of headers
to be included in the request. These are passed through a new parameter
in `client.Do()`.

In the future, to support dynamic headers, we can modify our codegen
template to add those fields to the headers map before calling the
client.

Alternatives considered:
* **Adding headers in request objects directly**: this would not require
any change in the client interface. Instead, we would introduce a new
annotation for headers, much like what we do for query parameters, and
reflectively scan the fields of each request object. We could do this,
but for headers that have a fixed value (like `Content-Type` and
`Accept`), these fields would essentially be private final fields that
we could iterate through. Users can never interact with these fields, so
it seems a bit unusual to have them in the request structure, a public
interface.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied