Support io.ReadCloser for request and response bodies (#590)
## Changes
This PR adds support for non-application/json requests and responses.
This PR is divided into two parts:

**Support setting headers per request**. The `Do` method of
DatabricksClient is expanded to accept one new parameter, `headers
map[string]string`, and all callers are expected to pass a map of
headers to be included in the request. As the allowed content types for
requests and responses are known from the OpenAPI specification, callers
must construct the request header map in code generation and pass it to
the client. The default "Content-Type: application/json" header is
removed, as each request should specify its own Content-Type and Accept
headers. This is done in
https://github.com/databricks/databricks-sdk-go/pull/589.

**Add support for streaming request and response bodies to support
non-application/json requests/responses**. Today, serialization of
requests is done in the client. This implies that ApiClient needs to be
able to serialize a request purely based on the parameters provided to
it via headers, the request body, etc. In this proposal, the
serialization is specified by the caller depending on whether the value
passed to the `request` and `response` parameters is an `io.ReadCloser`
or not.

One important caveat about streaming responses is that callers are
required to close the streams when they are done using them.
## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied