Reuse tokens after first call to Azure CLI (#605)
## Changes

The call to `oauth2.ReuseTokenSource` in `oauth_visitors.go` doesn't
include a token to start with so will always trigger a new call to
acquire a token.

We fix this by wrapping the `oauth2.TokenSource` with
`oauth2.ReuseTokenSource` from the caller, where we have a copy of a
valid token.

Calling `oauth2.ReuseTokenSource` twice is a no-op:
https://cs.opensource.google/go/x/oauth2/+/refs/tags/v0.11.0:oauth2.go;l=370

This commit also updates the logging to emit the info line only once we
know we can return from `Configure` successfully.

## Tests

- [x] `make test` passing
- [x] `make fmt` applied
- [ ] relevant integration tests applied