Fix case where retry of a request without body fails (#614)
## Changes

This problem was introduced in #592. It contains an assumption that a
`nil` body on the newly introduced `requestBody` struct cannot be `nil`,
yet it can be for requests without a body (HEAD, GET, and DELETE). If a
request of one of these types fails with a retriable error (e.g. a 429),
the code tries to reset the body, which in turn fails because it is nil.

The symptom is a GET request failing with: `cannot reset nil reader`.

Because a nil request body is valid, this change updates the expectation
accordingly.

## Tests

- [x] `make test` passing
- [x] `make fmt` applied
- [x] relevant integration tests applied