Adjust token expiry window on Azure
By default, the oauth2 library refreshes a token 10 seconds before it expires.
Azure Databricks rejects tokens that expire in 30 seconds or less.
This leaves a 20 second window where the SDK assumes the token is invalid while
the backend will reject it.

This change extends the token expiry window to 40 seconds to fix this.