Add details to APIError (#613)
## Changes
Databricks API can return error details in case of errors. In some
cases, users need to access such details to be able to solve the issue.
This is the case for the errors of type ErrorInfo used by the UC
catalog.
This PR adds the necessary fields to the APIError to Unmarshal the
ErrorDetails and to expose the details of type ErrorInfo to the users.

## Tests

- [X] `make test` passing
- [X] `make fmt` applied
- [x] relevant integration tests applied