Add support for template libraries (#635)
## Changes
This PR adds support in the codegen utility for template libraries.
These are templates that are reused across multiple stages of code
generation. For example, in the Java SDK, we need to compute the type of
a field/parameter/return value multiple times in different phases of the
SDK. This PR allows us to define these templates a single time and reuse
them across multiple files.

## Tests
Added to the Java SDK and regenerated the SDK.