Update OpenAPI Spec with Go SDK (#636)
## Changes
This PR bumps the version of the OpenAPI spec used in the Go SDK. To
support listing bare workspace IDs in the
AccountMetastoreAssignments.ListAll() method, a small tweak is made to
the return type in our template for paginated methods.

## Tests
Added an integration test to verify that the ListAll method works as
expected.