Prefix library paths with the target directory to construct absolute paths (#656)
## Changes
The OpenAPI generator tool for Databricks recently added support for
template libraries to have reusability across go templates in each SDK
repo across different generation scopes (e.g. Batch/Service/Type/etc.).
There is a bug that as a result causes generation to fail if CWD is not
the root of the repo being generated. This PR fixes this.

## Tests
Successfully regenerated the Java SDK with CWD in a different directory.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied