Listable resource iteration through Iterator (#543)
## Changes
A prototype of iterable functions for listable resources. The current Go
SDK diverges from the Java and Python SDKs in how list operations are
handled. The Go SDK synchronously exhaustively lists the resource (based
on the given request) and returns a slice of all entries, and "limit"
fields are interpreted as a global limit on the number of resources to
return (i.e. the length of the returned slice is never longer than
limit). In Python and Java, resource listing is done on demand by
returning an iterator/generator abstraction from the list API
immediately and only listing resources when requested. In these SDKs,
the limit field is passed to the API and is only interpreted as the page
size, so the number of returned resources is not bounded (though the
number returned per REST API call is).

Besides this, in Go, there is not currently a common iterator
abstraction. Even in the standard library, there are several different
conventions used for iterator-like operations, such as `(Scanner)
Scan()` (returning nil when scanning is complete). Without a common
iterator interface, range expressions are not supported out of the box.
See https://github.com/golang/go/discussions/54245 and
https://github.com/golang/go/discussions/56413 for in-depth discussions
about a proposed generic iterator interface in Go.

In light of all of this, this PR is a proposal for unifying the
approaches taken between the Go and Python/Java SDKs to return an
`Iterator` interface from listable APIs. This interface exposes
`Next(context.Context) (T, error)` and `HasNext(context.Context)`
methods that make it simple to iterate through the list of resources.
This can be used for all endpoints using resource listing, including
paginated and non-paginated endpoints alike. This reflects the
implementation of the Java SDK and Java iterators, which also have
`next()` and `hasNext()` methods.

`Next` returns the next object in the list, or if there are no more
objects in the list, it returns the zero value and
`listing.ErrNoMoreValues`. `HasNext` returns true if and only if there
are more items to process, with one caveat: for some APIs (specifically
offset-based APIs that don't include any sentinel value), it is not
possible to determine whether there is a next item in the list response
without actually making an API call. Because of this, `HasNext` must
take a `context.Context` parameter. In case the API call fails,
`HasNext` will return `true`, and the error can be retrieved by calling
`Next`.

If an error occurs while making an API call, that error is cached and no
further results are returned or fetched. In this case, the user is
expected to call the List API once more to get a fresh iterator.

## Other possible idioms
[`runners.CallersFrames`](https://pkg.go.dev/runtime#CallersFrames)
returns a pointer to a named type `*Frames` with a method `Next()
(*Frames, bool)`. The boolean parameter indicates if there is another
element available after the returned element. This mechanism does not
support returning errors, though the error could be added as a third
parameter.

[`bufio.Scanner`](https://pkg.go.dev/bufio#Scanner.Scan) provides 3
methods: `Scan() bool` indicating whether the scan has stopped (i.e. has
reached EOF), `Text() string` the element returned by the last call to
`Scan()`, and `Err() error` the error that occurred during scanning
(unless the end of the file was reached). This is a slightly more
cumbersome API requiring calls to all three methods separately in each
loop, but `Scan()` can be used in a range expression.

[`database/sql.Rows`](https://pkg.go.dev/database/sql#Rows.Next)
provides a `Next() bool` to advance the result set to the next row and a
`Scan(dest ...any) error` which scans the row into provided variables,
returning an error if one occurred. This is suitable for a very generic
endpoint (such as listing the results of a SQL query), but for most REST
API calls where the type of the result is well known, it is less natural
than simply returning a value.

## What to do with the existing List API?
This raises a question of how to handle the existing *All methods. We
can leave these *All methods in the SDK for convenience and because we
have no need to break them now. However, in some sense, they increase
the complexity of this project by having two ways to use an iterator,
the first being `listing.ToSlice(w.Service.List(ctx))` or
`listing.ToSliceN(w.Service.List(ctx), n)`, the second being
`w.Service.ListAll(ctx)`. The `ListAll` methods have an additional quirk
that they use the Limit field in the request as the global limit for
responses, which is different from our how other SDKs work. I think we
can leave this here safely for now, but when the time comes for v1.0.0,
we can remove these methods and encourage users to use the helper
methods from the `listing` package directly. I don't feel the urgency to
even mark these as deprecated at the moment, so we'll leave them as is.

## Tests
- [x] `iteratorImpl` and `dedupeIteratorImpl` are unit tested.