Update SDK to latest OpenAPI Spec (#685)
## Changes
<!-- Summary of your changes that are easy to understand -->
Generate SDK on latest OpenAPI specification

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
There were following nightly test failures:
- `TestAccJobsApiFullIntegration`: timeout 
- `RecipientActivationNoTranspile`: "Activation nonce does not exist" --
this is a known issue and has been happening on main branch as well so
not related to the changes.
- `CatalogWorkspaceBindings` (gcp-prod-ucws): This probably has to do
with the recent infra updates on gcp ucws. It doesn't seem related to
updated changes. Triggered another run to see if this is happening on
main branch as well.