Skip recipients tests in Azure (#692)
## Changes
The recipient activation flow in the SDK doesn't work properly today
because the API requires that this request is sent to the host specified
in the activation URL of the token generated when creating the
recipient. Currently, the SDK only supports using the host from the
static configuration. This does work in AWS and GCP today but purely by
coincidence, so for now we will stop running these tests in Azure.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied

Signed-off-by: Tomo Cesnik <tcesnik@veza.com>