Handle double slash in Files API (#712)
## Changes
Copy the URI-cleaning logic from the Python & Java SDKs to the Go SDK
removing empty path segments from requests to the Files API. This is
more conservative than what was done before, but it brings consistency
between the Go SDK and the other SDKs.

## Tests
Unit test for this case.
Files API test passes.