Support overriding DatabricksEnvironment (#723)
## Changes
In testing scenarios when using a mock HTTP server, environment
resolution will always return the AWS Prod environment, as that is the
fallback. In order to mock out other environments, this PR exposes the
fields of DatabricksEnvironment and the available AzureEnvironments,
allowing users to configure custom environments in tests. Under test, if
specified, the first value of the DatabricksEnvironments field is
returned from the `Environment()` method.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied