Improve debuggability of unexpected errors (#744)
## Changes
Issues like
https://github.com/databricks/terraform-provider-databricks/issues/3031
are difficult for users to debug when Databricks APIs respond with
unexpected response bodies. This is often a bug in either the API
(returning a non-JSON error message, for example) or from the caller
(e.g. calling update but not specifying the ID of the resource to
update), though it could also be due to a bug in the SDK itself.

This PR improves the debuggability of errors returned from the SDK due
to malformed responses or other failures to process the response by
including the request log in the error, along with helpful text that
points the user to report the issue at the Go SDK so it can be triaged
and fixed.

The returned error's `Error()` method will produce a string like this:

unexpected error handling request: invalid character '<' looking for
beginning of value. This is likely a bug in the Databricks SDK for Go or
the underlying REST API. Please report this issue with the following
debugging information to the SDK issue tracker at
https://github.com/databricks/databricks-sdk-go/issues. Request log:
```
GET /a
> * Host: 
> * Accept: application/json
> * Authorization: REDACTED
> * User-Agent: unknown/0.0.0 databricks-sdk-go/0.28.0 go/1.21.3 os/darwin auth/pat
< HTTP/2.0 Internal Server Error
< <html><body>hello</body></html>
```

Note that the markdown ticks are present in the request log itself for
ease of copy-pasting into Github's new issue UI.

## Tests
- [x] New unit tests cover RoundTripStringer.