Handle empty response for errors (#756)
## Changes
Credential deletion fails with 409 and no response body if the
credential is in use with a running workspace. This leads to some race
conditions because workspace deletion is asynchronous, so attempting to
delete a credential while a workspace is running will ultimately fail.
This PR tolerates empty error bodies. Note that `Unwrap()` for APIError
will delegate this to the appropriate status code-based error, so users
can still use `errors.Is` to examine errors with no body.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied