Add utility to retry on specific errors (#757)
## Changes
The SDK may return errors that users want to retry. For instance, a user
that makes many SCIM operations in parallel and some fail with a
Conflict may want to retry those failed requests. To simplify this
process for users, this PR introduces a Retrier abstraction that allows
users to configure the timeout for retries, a set of errors that should
be retried, and/or a custom callback that can check whether a particular
error should be retried. Example from this PR:

```go
	t.Cleanup(func() {
		err := retries.New[struct{}](retries.OnErrors(apierr.ErrResourceConflict)).Wait(ctx, func(ctx context.Context) error {
			return a.Credentials.DeleteByCredentialsId(ctx, role.CredentialsId)
		})
		require.NoError(t, err)
	})
```

I have used this to retry credential deletion in TestMwsAccWorkspaces.

## Tests
- [x] Lots of unit tests.